Fix 404 error message when request cannot be found
Fix imports

Fix again the imports

Move functions around to avoid circular imports
