+++ src/python/WMCore/RequestManager/RequestDB/Interface/Admin/RequestManagement.py
@@ -6,8 +6,7 @@ API for administering requests in the database
 """
 
 import WMCore.RequestManager.RequestDB.Connection as DBConnect
-from cherrypy import HTTPError
-
+from WMCore.RequestManager.RequestDB.Interface.Request.GetRequest import requestID
 
 
 def deleteRequest(requestName):
@@ -15,14 +14,9 @@ def deleteRequest(requestName):
     _deleteRequest_
 
     delete the request from the database by its RequestName
-
     """
-    factory = DBConnect.getConnection()
-    finder =  factory(classname="Request.FindByName")
-    reqId = finder.execute(requestName)
-    if reqId == None:
-        raise HTTPError(404, 'Given requestName not found: %s' % requestName)
+    reqId = requestID(requestName)
 
     factory = DBConnect.getConnection()
-    deleteReq = factory(classname = "Request.Delete")
+    deleteReq = factory(classname="Request.Delete")
     deleteReq.execute(reqId)
